Initial commit of all files related to Analyzing output generated by Klee
Contained in this commit is a Parser for Klee log output (generated by --use-query-log=solver:smt),
a factorizer for bv constraints, and a renamer for bv constraints.  The factorizer has a series of
test cases and has been well tested.  The renamer is unsafe (see README.md) although a little work
should bring that into line.  Finally, the factorizer has a series of options to provide different
levels of factorization capabilities (again, see README.md).